Add automated iTerm2 configuration
Now the settings of iTerm and profiles can be set. The main settings
will need to be changed rarely. Most of the stuff happens in the
profiles. A profile is defined that can be copied to a folder from which
iTerm reads "dynamic" profiles. That profile is known as "default"
setting in the general iTerm settings.